refactor : 소켓 연결 코드 모듈화
